+++ reana-ui/src/selectors.js
@@ -10,6 +10,9 @@
 
 import { USER_ERROR } from "./errors";
 
+// Config
+export const getConfig = state => state.config;
+
 // Auth
 export const isSignedIn = state => !!state.auth.email;
 export const getUserEmail = state => state.auth.email;
@@ -22,7 +25,6 @@ export const getReanaTokenStatus = state => state.auth.reanaToken.status;
 export const loadingTokenStatus = state => state.auth.reanaToken.loading;
 export const getReanaTokenRequestedAt = state =>
   state.auth.reanaToken.requestedAt;
-export const getAnnouncement = state => state.auth.announcement;
 
 // Workflows
 export const loadingWorkflows = state => state.workflows.loadingWorkflows;